+++ frontend/src/Components/InternationalBestSeller/InternationalbestSeller.jsx
@@ -8,6 +8,10 @@ import Modal from '../Modal';
 function InternationalBestSeller() {
   const [show, setShow] = useState(false);
   const [item, setItem] = useState();
+  const handleSeeDetails = (item) => {
+    setShow(true);
+    setItem(item);
+  };
   const { data: bookData, loading, error } = useAPI("international");
   if (loading) return <p>Loading...</p>;
   if (error) return <p>Error: {error.message}</p>;
@@ -22,7 +26,7 @@ function InternationalBestSeller() {
                 <img className="bestSellerPic" src={item.volumeInfo.imageLinks?.thumbnail} alt="No pic" />
                 <p className="title">{item.volumeInfo.title}</p>
                 <p className="author">{item.volumeInfo.authors?.join(', ')}</p>
-                <button className='seeDetails' onClick={()=>{setShow(true);setItem(item)}}>See details</button>
+                <button className='seeDetails'  onClick={() => handleSeeDetails(item)}>See details</button>
               </div>
           ))}
         </Slider>
