+++ frontend/src/Components/Navbar/navbar.scss
@@ -55,9 +55,7 @@ $menu-hover-color: #555;
     color: black;
     transition: background-color 0.3s ease;
     font-size: 14px;
-    .fa-user {
-        margin-right: 5px;
-    }
+   
 
     &:hover {
         background-color: $hover-bg-color;
@@ -173,6 +171,13 @@ $menu-hover-color: #555;
         color: $menu-hover-color;
         cursor: pointer;
     }
+}
+.linkTo{
+    text-decoration: none;
+}
+.fa-icon-navbar {
+    margin-right: 5px;
+    color: red;
 }
 
 @media (max-width: 1300px) {
