+++ frontend/src/pages/LoginPage/loginpage.css
@@ -1,0 +1,11 @@
+.passwordHolder{
+    all: unset;
+    position: relative;
+}
+.showPassword{
+    position: absolute;
+    top: 10px;
+    right: 10px;
+    color: white;
+    font-size: 14px;
+}
